backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.4
django-environ==0.11.2
django-ninja==1.3.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
